Solved day 8 part 2
My real part 2 solution in the previous commit was too slow to finish,
but I realized I could just calculate the number of steps for each
starting node individually, and then the least common multiple of the
step counts would be the answer. I used this commit to print the step
counts, then calculated the LCM in Wolfram Alpha, and it worked!